migrated getServerSideProps
