+++ packages/nextjs/components/scaffold-eth/Address.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useEffect, useState } from "react";
 import Link from "next/link";
 import Blockies from "react-blockies";
