+++ packages/nextjs/components/scaffold-eth/Balance.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useAccountBalance } from "~~/hooks/scaffold-eth";
 import { getTargetNetwork } from "~~/utils/scaffold-eth";
