+++ packages/nextjs/components/scaffold-eth/index.tsx
@@ -1,3 +1,4 @@
+"use client";
 export * from "./Address";
 export * from "./Balance";
 export * from "./BlockieAvatar";
